+++ _example/live-prefix/main.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 
 	prompt "github.com/elk-language/go-prompt"
+	istrings "github.com/elk-language/go-prompt/strings"
 )
 
 var LivePrefix string = ">>> "
@@ -17,14 +18,18 @@ func executor(in string) {
 	LivePrefix = in + "> "
 }
 
-func completer(in prompt.Document) []prompt.Suggest {
+func completer(in prompt.Document) ([]prompt.Suggest, istrings.RuneNumber, istrings.RuneNumber) {
+	endIndex := in.CurrentRuneIndex()
+	w := in.GetWordBeforeCursor()
+	startIndex := endIndex - istrings.RuneCount(w)
+
 	s := []prompt.Suggest{
 		{Text: "users", Description: "Store the username and age"},
 		{Text: "articles", Description: "Store the article text posted by user"},
 		{Text: "comments", Description: "Store the text commented to articles"},
 		{Text: "groups", Description: "Combine users with specific rules"},
 	}
-	return prompt.FilterHasPrefix(s, in.GetWordBeforeCursor(), true)
+	return prompt.FilterHasPrefix(s, w, true), startIndex, endIndex
 }
 
 func changeLivePrefix() string {